Split tests to standalone3
