+++ python/ray/serve/BUILD
@@ -296,6 +296,14 @@ py_test(
 )
 
 py_test(
+    name = "test_standalone3",
+    size = "medium",
+    srcs = serve_tests_srcs,
+    tags = ["exclusive", "team:serve"],
+    deps = [":serve_lib"],
+)
+
+py_test(
     name = "test_cluster",
     size = "medium",
     srcs = serve_tests_srcs,
